backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
psycopg2
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
python-decouple
requests
loguru
requests-middleware==0.1.2
django-apscheduler==0.5.1
apscheduler==3.2.0
emoji
textblob
wordcloud
# spacy
# spacy-entity-linker
networkx
matplotlib
spicy
